moving antoan's models into place
